pyproject.toml
[project]
name = "vector2mcap"
version = "0.1.0"
description = "Convert Vector JSONL telemetry files to MCAP format using protobuf serialization"
readme = "README.md"
authors = [
    { name = "Noah Bresler", email = "noahbresler2002@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "mcap-protobuf-support>=0.5.3",
    "protobuf>=6.31.1",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
]

[project.scripts]
vector2mcap = "vector2mcap.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
